requirements.txt
streamlit==1.24.0
transformers==4.35.2
torch==2.1.0
Pillow==10.0.0
opencv-python==4.8.1.78
pandas==2.1.1
numpy==1.26.1
baidu-aip==4.18.11
requests==2.31.0
